(GH-118) Rename to Cake.Issues.PullRequests.AzureDevOps
