+++ README.md
@@ -1,33 +1,32 @@
-# TFS / Azure DevOps support for Cake Issues Addin
+# Azure DevOps support for Cake Issues Addin
 
 This addin for the Cake Build Automation System allows you to write issues found using any code
-analyzer or linter to Team Foundation Server or Azure DevOps pull requests using
-the [Cake Issues addin](https://github.com/cake-contrib/Cake.Issues).
+analyzer or linter to Azure DevOps pull requests using the [Cake Issues addin](https://github.com/cake-contrib/Cake.Issues).
 
 For more information about this add-in see the [Cake.Issues website](https://cakeissues.net)
 and for general information about the Cake build automation system see the [Cake website](http://cakebuild.net).
 
-[![License](http://img.shields.io/:license-mit-blue.svg)](https://github.com/cake-contrib/Cake.Issues.PullRequests.Tfs/blob/feature/build/LICENSE)
+[![License](http://img.shields.io/:license-mit-blue.svg)](https://github.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/blob/feature/build/LICENSE)
 
 ## Information
 
 | | Stable | Pre-release |
 |:--:|:--:|:--:|
-|GitHub Release|-|[![GitHub release](https://img.shields.io/github/release/cake-contrib/Cake.Issues.PullRequests.Tfs.svg)](https://github.com/cake-contrib/Cake.Issues.PullRequests.Tfs/releases/latest)|
-|NuGet|[![NuGet](https://img.shields.io/nuget/v/Cake.Issues.PullRequests.Tfs.svg)](https://www.nuget.org/packages/Cake.Issues.PullRequests.Tfs)|[![NuGet](https://img.shields.io/nuget/vpre/Cake.Issues.PullRequests.Tfs.svg)](https://www.nuget.org/packages/Cake.Issues.PullRequests.Tfs)|
+|GitHub Release|-|[![GitHub release](https://img.shields.io/github/release/cake-contrib/Cake.Issues.PullRequests.AzureDevOps.svg)](https://github.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/releases/latest)|
+|NuGet|[![NuGet](https://img.shields.io/nuget/v/Cake.Issues.PullRequests.AzureDevOps.svg)](https://www.nuget.org/packages/Cake.Issues.PullRequests.AzureDevOps)|[![NuGet](https://img.shields.io/nuget/vpre/Cake.Issues.PullRequests.AzureDevOps.svg)](https://www.nuget.org/packages/Cake.Issues.PullRequests.AzureDevOps)|
 
 ## Build Status
 
 | | Develop | Master |
 |:--:|:--:|:--:|
-|AppVeyor Windows|[![Build status](https://ci.appveyor.com/api/projects/status/s8t8j2gh6icynkod/branch/develop?svg=true)](https://ci.appveyor.com/project/cakecontrib/cake-issues-pullrequests-tfs/branch/develop)|[![Build status](https://ci.appveyor.com/api/projects/status/s8t8j2gh6icynkod/branch/master?svg=true)](https://ci.appveyor.com/project/cakecontrib/cake-issues-pullrequests-tfs/branch/master)|
-|Azure DevOps Windows|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.Tfs?branchName=develop&jobName=Windows)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_build/latest?definitionId=8&branchName=develop)|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.Tfs?branchName=master&jobName=Windows)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_build/latest?definitionId=8&branchName=master)|
-|Azure DevOps macOS|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.Tfs?branchName=develop&jobName=macOS)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_build/latest?definitionId=8&branchName=develop)|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.Tfs?branchName=master&jobName=macOS)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_build/latest?definitionId=8&branchName=master)|
-|Azure DevOps Ubuntu|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.Tfs?branchName=develop&jobName=Ubuntu)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_build/latest?definitionId=8&branchName=develop)|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.Tfs?branchName=master&jobName=Ubuntu)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.Tfs/_build/latest?definitionId=8&branchName=master)|
+|AppVeyor Windows|[![Build status](https://ci.appveyor.com/api/projects/status/6t6m39enwkvkhk80/branch/develop?svg=true)](https://ci.appveyor.com/project/cakecontrib/cake-issues-pullrequests-azuredevops/branch/develop)|[![Build status](https://ci.appveyor.com/api/projects/status/6t6m39enwkvkhk80/branch/master?svg=true)](https://ci.appveyor.com/project/cakecontrib/cake-issues-pullrequests-azuredevops/branch/master)|
+|Azure DevOps Windows|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.AzureDevOps?branchName=develop&jobName=Windows)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_build/latest?definitionId=8&branchName=develop)|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.AzureDevOps?branchName=master&jobName=Windows)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_build/latest?definitionId=8&branchName=master)|
+|Azure DevOps macOS|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.AzureDevOps?branchName=develop&jobName=macOS)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_build/latest?definitionId=8&branchName=develop)|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.AzureDevOps?branchName=master&jobName=macOS)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_build/latest?definitionId=8&branchName=master)|
+|Azure DevOps Ubuntu|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.AzureDevOps?branchName=develop&jobName=Ubuntu)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_build/latest?definitionId=8&branchName=develop)|[![Build Status](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_apis/build/status/cake-contrib.Cake.Issues.PullRequests.AzureDevOps?branchName=master&jobName=Ubuntu)](https://dev.azure.com/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/_build/latest?definitionId=8&branchName=master)|
 
 ## Code Coverage
 
-[![Coverage Status](https://coveralls.io/repos/github/cake-contrib/Cake.Issues.PullRequests.Tfs/badge.svg?branch=develop)](https://coveralls.io/github/cake-contrib/Cake.Issues.PullRequests.Tfs?branch=develop)
+[![Coverage Status](https://coveralls.io/repos/github/cake-contrib/Cake.Issues.PullRequests.AzureDevOps/badge.svg?branch=develop)](https://coveralls.io/github/cake-contrib/Cake.Issues.PullRequests.AzureDevOps?branch=develop)
 
 ## Quick Links
