+++ src/Cake.Issues.PullRequests.AzureDevOps.Tests/Cake.Issues.PullRequests.AzureDevOps.Tests.csproj
@@ -1,0 +1,35 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netcoreapp2.0</TargetFrameworks>
+    <IsPackable>false</IsPackable>
+    <Product>Cake.Issues.PullRequests.AzureDevOps</Product>
+    <Copyright>Copyright © BBT Software AG and contributors</Copyright>
+    <Description>Tests for the Cake.Issues.PullRequests.AzureDevOps addin</Description>
+    <Authors>BBT Software AG and contributors</Authors>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <CodeAnalysisRuleSet>..\Cake.Issues.PullRequests.AzureDevOps.Tests.ruleset</CodeAnalysisRuleSet>
+    <LangVersion>latest</LangVersion>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Cake.Core" Version="0.33.0" />
+    <PackageReference Include="Cake.Issues" Version="0.7.0" />
+    <PackageReference Include="Cake.Issues.PullRequests" Version="0.7.0" />
+    <PackageReference Include="Cake.Issues.Testing" Version="0.7.0" />
+    <PackageReference Include="Cake.Testing" Version="0.33.0" />
+    <PackageReference Include="Cake.AzureDevOps" Version="0.4.0-beta0001" />
+    <PackageReference Include="NSubstitute" Version="4.2.1" />
+    <PackageReference Include="Shouldly" Version="3.0.2" />
+    <PackageReference Include="StyleCop.Analyzers" Version="1.1.118" />
+    <PackageReference Include="xunit" Version="2.4.1" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Cake.Issues.PullRequests.AzureDevOps\Cake.Issues.PullRequests.AzureDevOps.csproj" />
+  </ItemGroup>
+
+</Project>
