+++ src/Cake.Issues.PullRequests.AzureDevOps.sln.DotSettings
