+++ src/Cake.Issues.PullRequests.AzureDevOps/AzureDevOpsPullRequestSystemAliases.cs
@@ -1,0 +1,12 @@
+﻿namespace Cake.Issues.PullRequests.AzureDevOps
+{
+    using Cake.Core.Annotations;
+
+    /// <summary>
+    /// Contains functionality related to writing code analysis issues to Azure DevOps pull requests.
+    /// </summary>
+    [CakeAliasCategory(IssuesAliasConstants.MainCakeAliasCategory)]
+    public static partial class AzureDevOpsPullRequestSystemAliases
+    {
+    }
+}
