--- src/Cake.Issues.PullRequests.Tfs.Tests/Cake.Issues.PullRequests.Tfs.Tests.csproj
@@ -1,35 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp2.0</TargetFrameworks>
-    <IsPackable>false</IsPackable>
-    <Product>Cake.Issues.PullRequests.Tfs</Product>
-    <Copyright>Copyright © BBT Software AG and contributors</Copyright>
-    <Description>Tests for the Cake.Issues.PullRequests.Tfs addin</Description>
-    <Authors>BBT Software AG and contributors</Authors>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <CodeAnalysisRuleSet>..\Cake.Issues.PullRequests.Tfs.Tests.ruleset</CodeAnalysisRuleSet>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Cake.Core" Version="0.33.0" />
-    <PackageReference Include="Cake.Issues" Version="0.7.0" />
-    <PackageReference Include="Cake.Issues.PullRequests" Version="0.7.0" />
-    <PackageReference Include="Cake.Issues.Testing" Version="0.7.0" />
-    <PackageReference Include="Cake.Testing" Version="0.33.0" />
-    <PackageReference Include="Cake.Tfs" Version="0.3.1" />
-    <PackageReference Include="NSubstitute" Version="4.2.1" />
-    <PackageReference Include="Shouldly" Version="3.0.2" />
-    <PackageReference Include="StyleCop.Analyzers" Version="1.1.118" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Cake.Issues.PullRequests.Tfs\Cake.Issues.PullRequests.Tfs.csproj" />
-  </ItemGroup>
-
-</Project>
