--- src/Cake.Issues.PullRequests.Tfs.sln
@@ -1,36 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28922.388
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cake.Issues.PullRequests.Tfs", "Cake.Issues.PullRequests.Tfs\Cake.Issues.PullRequests.Tfs.csproj", "{1D046545-035B-4D7E-98C8-8AE6F5D0D289}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cake.Issues.PullRequests.Tfs.Tests", "Cake.Issues.PullRequests.Tfs.Tests\Cake.Issues.PullRequests.Tfs.Tests.csproj", "{E938CD33-F1C7-45ED-A12C-7F18934AD131}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{AAD732D5-4A2D-4A67-B571-7CD931983393}"
-	ProjectSection(SolutionItems) = preProject
-		..\recipe.cake = ..\recipe.cake
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1D046545-035B-4D7E-98C8-8AE6F5D0D289}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D046545-035B-4D7E-98C8-8AE6F5D0D289}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D046545-035B-4D7E-98C8-8AE6F5D0D289}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D046545-035B-4D7E-98C8-8AE6F5D0D289}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E938CD33-F1C7-45ED-A12C-7F18934AD131}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E938CD33-F1C7-45ED-A12C-7F18934AD131}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E938CD33-F1C7-45ED-A12C-7F18934AD131}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E938CD33-F1C7-45ED-A12C-7F18934AD131}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1CAC9A70-DDF8-467C-8CA5-F65DED3BEF1C}
-	EndGlobalSection
-EndGlobal
