--- src/Cake.Issues.PullRequests.Tfs/Properties/AssemblyInfo.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("1d046545-035b-4d7e-98c8-8ae6f5d0d289")]
-
-[assembly: CLSCompliant(true)]
-[assembly: InternalsVisibleTo("Cake.Issues.PullRequests.Tfs.Tests")]
-[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Required for NSubstitute
