--- src/Cake.Issues.PullRequests.Tfs/TfsPullRequestSystemAliases.cs
@@ -1,13 +1,0 @@
-﻿namespace Cake.Issues.PullRequests.Tfs
-{
-    using Cake.Core.Annotations;
-
-    /// <summary>
-    /// Contains functionality related to writing code analysis issues to Team Foundation Server or
-    /// Azure DevOps pull requests.
-    /// </summary>
-    [CakeAliasCategory(IssuesAliasConstants.MainCakeAliasCategory)]
-    public static partial class TfsPullRequestSystemAliases
-    {
-    }
-}
